Refactor Syzygy probe methods and constants for clarity
Updated `BitPosition` method calls to attribute access for consistency and improved readability. Refactored arguments in probe methods and introduced new constants in `SyzygyConstants` to aid tablebase probing. Added initial scaffolding for `probe_root` and detailed DTZ/WDL probing logic.

+++ board/src/main/java/net/chesstango/board/representations/syzygy/EncInfo.java
@@ -5,7 +5,7 @@ package net.chesstango.board.representations.syzygy;
  */
 class EncInfo {
     PairsData precomp;
-    long[] factor = new long[SyzygyConstants.TB_PIECES];
+    int[] factor = new int[SyzygyConstants.TB_PIECES];
     byte[] pieces = new byte[SyzygyConstants.TB_PIECES];
     byte[] norm = new byte[SyzygyConstants.TB_PIECES];
 }
